libs/validation-constraints/src/lib/constraints/is-isbn.spec.ts
import {
    assertConstraintFailsForPredicate,
    assertConstraintSatisfiedForPredicate,
    buildInvalidCaseDescription,
    buildValidCaseDescription,
    SHOULD_RETURN_THE_EXPECTED_RESULT,
} from '../__tests__';
import { isISBN } from './is-isbn';

const validValues: string[] = ['0-395-36341-1'];

const invalidValues = [
    Math.PI,
    77.5,
    -77.5,
    3 / 8,
    true,
    false,
    'hello Mars!',
    null,
    undefined,
    {},
    [1999, 2000, 2001],
    Infinity,
    -Infinity,
    NaN,
];

describe('isInteger', () => {
    describe('when the value satisfies the constraint', () => {
        const assertConstraintSatisfied = assertConstraintSatisfiedForPredicate(isISBN);

        validValues.forEach((validValue) => {
            describe(buildValidCaseDescription(validValue), () => {
                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                    assertConstraintSatisfied(validValue);
                });
            });
        });
    });

    describe('when the value fails the constraint', () => {
        const assertConstraintFailure = assertConstraintFailsForPredicate(isISBN);

        invalidValues.forEach((invalidValue) => {
            describe(buildInvalidCaseDescription(invalidValue), () => {
                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                    assertConstraintFailure(invalidValue);
                });
            });
        });
    });
});
